+++ src/main/java/io/kontur/eventapi/entity/PgSetting.java
@@ -1,0 +1,9 @@
+package io.kontur.eventapi.entity;
+
+import lombok.Data;
+
+@Data
+public class PgSetting {
+    private String name;
+    private String setting;
+}